Add tests for email notifications
This adds a step to sign the Email Alerts user in Integration / Staging
(which is the only user who can sign up for emails - see https://docs.publishing.service.gov.uk/repos/email-alert-api/receiving-emails-from-email-alert-api-in-integration-and-staging.html#in-integration)
to email alerts for the section where our test document will be
created.

After a change to the block is made, we use the Notify API to check if
an email was sent to that email address for that particular page.